WIP: Got Subscribe working.
Key was to get all the characteristics as part of the constructor and needed a seperate callback for winble.

Need lots of cleanup in both cpp and spheropy.py.
Note that trying to make some of the sync calls into async calls messed up finding some of the services. I think this probably has to do with calling from UI thread.

Tested with set led and roll.